restructured files and folders for reporting module, active and resolved
