+++ src/modules/reporting/HorizontalChartLegend.jsx
