+++ src/modules/reporting/InsightsTable.jsx
