+++ src/modules/reporting/RecentVotesTable.jsx
