+++ src/modules/reporting/ReportingInfo.jsx
