+++ src/modules/reporting/VotesSummaryDoughnutCharts.jsx
