+++ src/modules/reporting/active/ActiveReportingEmptyState.jsx
@@ -1,7 +1,7 @@
 import { RegularButton } from "@/components/UI/atoms/button/regular";
 import { Label } from "@/components/UI/atoms/label";
 import { useEffect, useState } from "react";
-import { ReportingDropdown } from "@/components/UI/molecules/reporting/reporting-dropdown";
+import { ReportingDropdown } from "@/src/modules/reporting/reporting-dropdown";
 import { useRouter } from "next/router";
 import { actions } from "@/src/config/cover/actions";
 import { getCoverImgSrc, getParsedKey } from "@/src/helpers/cover";
@@ -29,13 +29,13 @@ export const ActiveReportingEmptyState = () => {
   }
 
   return (
-    <div className="w-full flex flex-col items-center pt-20">
+    <div className="flex flex-col items-center w-full pt-20">
       <img
         src="/images/covers/empty-list-illustration.svg"
         alt="no data found"
         className="w-48 h-48"
       />
-      <p className="text-h5 text-404040 text-center mt-8 w-96 max-w-full">
+      <p className="max-w-full mt-8 text-center text-h5 text-404040 w-96">
         No known incident found for any cover product. If you believe a cover
         incident has occurred, earn rewards by reporting the incident.
       </p>
@@ -48,7 +48,7 @@ export const ActiveReportingEmptyState = () => {
           selected={selected}
           setSelected={setSelected}
           prefix={
-            <div className="w-8 h-8 p-1 mr-2 bg-DEEAF6 rounded-full">
+            <div className="w-8 h-8 p-1 mr-2 rounded-full bg-DEEAF6">
               <img src={getCoverImgSrc(selected)} alt={selected?.coverName} />
             </div>
           }
