+++ src/modules/reporting/active/active.jsx
@@ -3,64 +3,60 @@ import { NeutralButton } from "@/components/UI/atoms/button/neutral-button";
 import { Container } from "@/components/UI/atoms/container";
 import { Grid } from "@/components/UI/atoms/grid";
 import { SearchAndSortBar } from "@/components/UI/molecules/search-and-sort";
-import { ResolvedReportingCard } from "@/components/UI/organisms/reporting/ResolvedReportingCard";
-import { ReportStatus } from "@/src/config/constants";
-import { useCovers } from "@/src/context/Covers";
+import { ActiveReportingCard } from "@/src/modules/reporting/active/ActiveReportingCard";
+import { ActiveReportingEmptyState } from "@/src/modules/reporting/active/ActiveReportingEmptyState";
+import { useActiveReportings } from "@/src/hooks/useActiveReportings";
 import { getParsedKey } from "@/src/helpers/cover";
-import { useResolvedReportings } from "@/src/hooks/useResolvedReportings";
-import { useSearchResults } from "@/src/hooks/useSearchResults";
 import Link from "next/link";
+import { useSearchResults } from "@/src/hooks/useSearchResults";
+import { useCovers } from "@/src/context/Covers";
 import { sortData } from "@/utils/sorting";
 import { CardSkeleton } from "@/components/common/Skeleton/CardSkeleton";
 import { COVERS_PER_PAGE } from "@/src/config/constants";
 
-export const ReportingResolvedPage = () => {
-  const { data, loading, hasMore, handleShowMore } = useResolvedReportings();
+export const ReportingActivePage = () => {
+  const { data, loading, hasMore, handleShowMore } = useActiveReportings();
   const { getInfoByKey } = useCovers();
   const { searchValue, setSearchValue, filtered } = useSearchResults({
     list: data.incidentReports,
     filter: (item, term) => {
       const info = getInfoByKey(item.key);
+
       return info.projectName.toLowerCase().indexOf(term.toLowerCase()) > -1;
     },
   });
 
   const [sortType, setSortType] = useState({ name: "A-Z" });
 
-  const filteredResolvedCardInfo = filtered.map((item) => {
-    const resolvedCardInfo = getInfoByKey(item.key);
+  const filteredActiveCardInfo = filtered.map((item) => {
+    const activeCardInfo = getInfoByKey(item.key);
 
-    return { ...resolvedCardInfo, resolvedReporting: item };
+    return { ...activeCardInfo, activeReporting: item };
   });
 
   const searchHandler = (ev) => {
     setSearchValue(ev.target.value);
   };
 
-  const renderResolvedReportings = () => {
+  const renderActiveReportings = () => {
     const noData = data.incidentReports.length <= 0;
 
     if (!loading && !noData) {
       return (
         <Grid className="mb-24 mt-14">
-          {sortData(filteredResolvedCardInfo, sortType.name).map(
-            ({ resolvedReporting }) => {
-              const resolvedOn = resolvedReporting.emergencyResolved
-                ? resolvedReporting.emergencyResolveTransaction?.timestamp
-                : resolvedReporting.resolveTransaction?.timestamp;
-
+          {sortData(filteredActiveCardInfo, sortType.name).map(
+            ({ activeReporting }) => {
               return (
                 <Link
                   href={`/reporting/${getParsedKey(
-                    resolvedReporting.id.split("-")[0]
-                  )}/${resolvedReporting.id.split("-")[1]}/details`}
-                  key={resolvedReporting.id}
+                    activeReporting.id.split("-")[0]
+                  )}/${activeReporting.id.split("-")[1]}/details`}
+                  key={activeReporting.id}
                 >
                   <a className="rounded-3xl focus:outline-none focus-visible:ring-2 focus-visible:ring-4e7dd9">
-                    <ResolvedReportingCard
-                      coverKey={resolvedReporting.key}
-                      resolvedOn={resolvedOn}
-                      status={ReportStatus[resolvedReporting.status]}
+                    <ActiveReportingCard
+                      coverKey={activeReporting.key}
+                      incidentDate={activeReporting.incidentDate}
                     />
                   </a>
                 </Link>
@@ -70,15 +66,13 @@ export const ReportingResolvedPage = () => {
         </Grid>
       );
     } else if (!loading && noData) {
-      return <p className="text-center">No data found</p>;
+      return <ActiveReportingEmptyState />;
     }
 
     return (
       <Grid className="mb-24 mt-14">
         <CardSkeleton
           numberOfCards={data.incidentReports.length || COVERS_PER_PAGE}
-          subTitle={false}
-          lineContent={1}
         />
       </Grid>
     );
@@ -95,7 +89,7 @@ export const ReportingResolvedPage = () => {
         />
       </div>
 
-      {renderResolvedReportings()}
+      {renderActiveReportings()}
 
       {!loading && hasMore && (
         <NeutralButton
