+++ src/modules/reporting/resolved/ResolvedReportingCard.jsx
